feat: add configuration page
