+++ src/app/projects/[ownerType]/[org]/[project]/configure/page.js
@@ -1,0 +1,7 @@
+import ProjectConfigurationPage from "@/components/ProjectConfigurationPage"
+
+export default async function ProjectConfigurationPageServer({ params }) {
+  return (
+    <ProjectConfigurationPage />
+  )
+}
