+++ src/app/projects/[ownerType]/[org]/[project]/page.js
@@ -1,30 +1,8 @@
-import CyclePage from '@/components/CyclePage'
-import { prepareData } from './cycles/[cycle]/page'
-import { ProjectDetailsProvider } from '@/contexts/ProjectDetails'
+import { redirect } from 'next/navigation'
+import { fetchProject } from './layout'
 
 export default async function ProjectPage({ params }) {
-  const {
-    project,
-    visibleCycle,
-    previousCycle,
-    nextCycle,
-    inCycle,
-    availablePitches = [],
-    availableBets = [],
-    availableScopes = []
-  }  = await prepareData(params)
-
-  return (
-    <ProjectDetailsProvider value={project}>
-      <CyclePage
-        visibleCycle={visibleCycle}
-        previousCycle={previousCycle}
-        nextCycle={nextCycle}
-        inCycle={inCycle}
-        availablePitches={availablePitches}
-        availableBets={availableBets}
-        availableScopes={availableScopes}
-      />
-    </ProjectDetailsProvider>
-  )
+  const project = await fetchProject(params)
+  if (!project.currentCycle) return redirect(`/projects/${params.ownerType}/${params.org}/${params.project}/configure`)
+  return redirect(`/projects/${params.ownerType}/${params.org}/${params.project}/cycles/${project.currentCycle.id}`)
 }
