+++ src/app/projects/page.js
@@ -124,8 +124,6 @@ export default async function Projects() {
   ).flat()
   const allProjects = [userProjects, orgProjects].flat()
 
-  console.log(JSON.stringify(allProjects, null, 2))
-
   return (
     <>
       <Head>
