+++ src/components/Header.js
@@ -2,20 +2,15 @@ import Link from 'next/link'
 import { useProjectDetails } from '@/contexts/ProjectDetails'
 
 export default function Header() {
-  const { title, number, org, ownerType, githubUrl } = useProjectDetails()
+  const { title, number, org, ownerType, url } = useProjectDetails()
 
   return (
     <div className="lg:flex lg:justify-between">
-      <div className="max-w-xl">
-        <Link href={`/projects/${ownerType}/${org}/${number}`}>
-          <h2 className="text-4xl leading-10 font-extrabold text-gray-900 sm:text-5xl sm:leading-none sm:tracking-tight lg:text-6xl">
-            <span className="shapeup-animated-gradient">{title}</span>
-          </h2>
-        </Link>
-        <p className="mt-5 text-xl leading-7 text-gray-500 prose prose-pink">
-          Visualize the progress of your work.
-        </p>
-      </div>
+      <Link href={`/projects/${ownerType}/${org}/${number}`}>
+        <h2 className="text-4xl leading-10 font-extrabold text-gray-900 sm:text-5xl sm:leading-none sm:tracking-tight lg:text-6xl">
+          <span className="shapeup-animated-gradient">{title}</span>
+        </h2>
+      </Link>
       <div className="flex flex-col space-y-4">
         <a href={`/projects/${ownerType}/${org}/${number}/rss.xml`} target="_blank" title="RSS Feed" className="block translate-y-1 text-gray-400 hover:text-gray-800">
           <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
@@ -23,7 +18,7 @@ export default function Header() {
           </svg>
           <span className="sr-only">RSS Feed</span>
         </a>
-        <a href={githubUrl} title="GitHub Project" target="_blank" className="block text-gray-400 hover:text-gray-800">
+        <a href={url} title="GitHub Project" target="_blank" className="block text-gray-400 hover:text-gray-800">
           <span className="sr-only">Github</span>
           <svg className="h-6 w-6" aria-hidden="true" fill="currentColor" viewBox="0 0 24 24">
             <path fillRule="evenodd" d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z" clipRule="evenodd" />
